+++ CHANGES.md
@@ -4,6 +4,8 @@
 - support `FormPdoAuthentication` and make it the default
 - deprecate `FormAuthentcation`, new deploys will use `FormPdoAuthentication` 
   by default
+- implement support for changing passwords by users when using 
+  `FormPdoAuthentication`
 
 ## 1.4.9 (2018-02-06)
 - set default for refresh token expiry to 180 days (instead of 6 months to sync